Rename executive_summary interfaces module.
