+++ xcp_d/interfaces/__init__.py
@@ -6,7 +6,7 @@ from xcp_d.interfaces import (
     bids,
     c3,
     connectivity,
-    executive_summary,
+    execsummary,
     filtering,
     nilearn,
     plotting,
@@ -24,7 +24,7 @@ __all__ = [
     "c3",
     "connectivity",
     "filtering",
-    "executive_summary",
+    "execsummary",
     "nilearn",
     "prepostcleaning",
     "plotting",
